+++ mdns_plugin/lib/mdns_plugin.dart
@@ -2,6 +2,8 @@ import 'dart:async';
 import 'dart:convert';
 import 'package:flutter/services.dart';
 
+/////////////////////////////////////////////////////////////////////
+
 abstract class MDNSPluginDelegate {
   void onDiscoveryStarted();
   void onDiscoveryStopped();
@@ -10,6 +12,8 @@ abstract class MDNSPluginDelegate {
   void onServiceUpdated(MDNSService service);
   void onServiceRemoved(MDNSService service);
 }
+
+/////////////////////////////////////////////////////////////////////
 
 class MDNSService {
   final Map map;
@@ -25,34 +29,59 @@ class MDNSService {
   String get serviceType => map["type"];
   int get port => map["port"];
   Map get txt => map["txt"];
+  List<String> get addresses {
+    var addresses = map["address"];
+
+    if (addresses is List<dynamic>) {
+      var address = List<String>();
+      addresses.forEach((value) {
+        if (value.length == 2 && value[0] is String) {
+          address.add(value[0]);
+        }
+      });
+      return address;
+    } else {
+      return [];
+    }
+  }
 
   // METHODS ////////////////////////////////////////////////////////
 
   static String toUTF8String(List<int> bytes) {
-    return Utf8Codec().decode(bytes);
+    if(bytes == null) {
+      return null;
+    } else {
+      return Utf8Codec().decode(bytes);
+    }
   }
 
   String toString() {
     var parts = "";
-    if(name != "") {
+    if (name != "") {
       parts = parts + "name='$name' ";
     }
-    if(serviceType != "") {
+    if (serviceType != "") {
       parts = parts + "serviceType='$serviceType' ";
-    }    
-    if(hostName != "" && port > 0) {
+    }
+    if (hostName != "" && port > 0) {
       parts = parts + "host='$hostName:$port' ";
     }
-    txt.forEach((k,v) {
+    if (addresses.length > 0) {
+      parts = parts + "addresses=$addresses ";
+    }
+    txt.forEach((k, v) {
       var vstr = toUTF8String(v);
       parts = parts + "$k='$vstr' ";
-    }); 
+    });
     return "<MDNSService>{ $parts}";
   }
 }
 
+/////////////////////////////////////////////////////////////////////
+
 class MDNSPlugin {
-  static const MethodChannel _methodChannel = const MethodChannel('mdns_plugin');
+  static const MethodChannel _methodChannel =
+      const MethodChannel('mdns_plugin');
   final EventChannel _eventChannel = const EventChannel('mdns_plugin_delegate');
   final MDNSPluginDelegate delegate;
 
@@ -60,8 +89,8 @@ class MDNSPlugin {
 
   MDNSPlugin(this.delegate) : assert(delegate != null) {
     _eventChannel.receiveBroadcastStream().listen((args) {
-      if(args is Map && args.containsKey("method")) {
-        switch(args["method"]) {
+      if (args is Map && args.containsKey("method")) {
+        switch (args["method"]) {
           case "onDiscoveryStarted":
             delegate.onDiscoveryStarted();
             break;
@@ -88,15 +117,17 @@ class MDNSPlugin {
   // METHODS ////////////////////////////////////////////////////////
 
   static Future<String> get platformVersion async {
-    final String version = await _methodChannel.invokeMethod('getPlatformVersion');
+    final String version =
+        await _methodChannel.invokeMethod('getPlatformVersion');
     return version;
   }
 
-  Future<void> startDiscovery(String serviceType,{String domain}) async {
-    return await _methodChannel.invokeMethod("startDiscovery",{ "serviceType": serviceType, "domain": domain });
+  Future<void> startDiscovery(String serviceType, {String domain}) async {
+    return await _methodChannel.invokeMethod(
+        "startDiscovery", {"serviceType": serviceType, "domain": domain});
   }
 
   Future<void> stopDiscovery() async {
-    return await _methodChannel.invokeMethod('stopDiscovery',{ });
+    return await _methodChannel.invokeMethod('stopDiscovery', {});
   }
 }